bin/exec() as independent sub module
