+++ src/bin/upload.ts
@@ -1,33 +1,14 @@
-import * as child_process from 'child_process';
 import {sequenceT} from 'fp-ts/lib/Apply';
 import {getMonoid} from 'fp-ts/lib/Array';
-import {left, right} from 'fp-ts/lib/Either';
-import {Task} from 'fp-ts/lib/Task';
 import {TaskEither, taskEither} from 'fp-ts/lib/TaskEither';
 import {Package} from './decoders';
+import {ExecOutput, exec} from './exec';
 import {Program} from './program';
 import {ReadPkg, readPkg} from './read-pkg';
 import {Trace, trace} from './trace';
 
 const sequenceTE = sequenceT(taskEither);
 const concatS = getMonoid<string>().concat;
-
-interface ExecOutput {
-  stdout: string;
-  stderr: string;
-}
-
-const exec = (cmd: string): TaskEither<Error, ExecOutput> =>
-  new TaskEither(
-    new Task(
-      () =>
-        new Promise(resolve =>
-          child_process.exec(cmd, {encoding: 'utf-8'}, (err, stdout, stderr) =>
-            err !== null ? resolve(left(err)) : resolve(right({stdout, stderr}))
-          )
-        )
-    )
-  );
 
 const prepareOpts = (pkg: Package, args: string[]): string =>
   concatS(
